The Autonomous Toy Car is a project whose purpose is to automate a toy car to allow it to drive by itself without crashing with any object by creating a system that will provide environment and position awareness to the robot based on sensor data (Inertial Measurement Unit, and Ultrasonic Sensor)
The Autonomous Toy Car Project aimed to automate a toy car from scratch, using an Arduino Uno R3, an Inertial Measurement Unit (IMU) -which measurers acceleration, magnetism and angles in x, y, z axis-, and an Ultrasonic sensor -which measures distance- to sense the environment and control the movement of the vehicle to allow the prototype avoid objects. The purpose of this project was to allow the toy car to autonomously drive without crashing with any object in the way by making smart decisions on where to move based on sensor data.
To accomplish this, the Arduino controller board used an algorithm that allowed it to decide whether an obstacle is too close (40-50 cm or 15.75-19.70 in) to the toy car based on data from the ultrasonic sensor. Then the algorithm decided to make the car turn right or left based on the greatest average distance sensed at each side of the car. To take it even further, the algorithm determined what angle is the best to turn to at either the left or the right side by using a combination of the IMU, ultrasonic sensor, and the angle of the servo motor1. Now, to make this possible, the Arduino board will change the direction of the current into each pair of motors in the right and left sides of the vehicle independently by using a motor driver. That way, when the algorithm decides there is an obstacle considerably close, it will enable the Arduino to stop both motors from working. Then when the program decides where to turn, the Arduino will activate the pair of motors at the opposite side of where the program decided the car should turn to move in the forward direction and the other pair of motors to move in the opposite direction. This will make the toy car rotate towards the desired direction unlike a regular car’s steering system but imitating the turning system of a war tank2.